+++ Simulation/Simulation.vcxproj
@@ -149,7 +149,7 @@
     <ClInclude Include="com_Shader.h" />
     <ClInclude Include="com_Sphere.h" />
     <ClInclude Include="Stage.h" />
-    <ClInclude Include="Texture.h" />
+    <ClInclude Include="com_Texture.h" />
     <ClInclude Include="Theatre.h" />
     <ClInclude Include="Transform.h" />
   </ItemGroup>
@@ -163,7 +163,7 @@
     <ClCompile Include="main.cpp" />
     <ClCompile Include="com_Sphere.cpp" />
     <ClCompile Include="Stage.cpp" />
-    <ClCompile Include="Texture.cpp" />
+    <ClCompile Include="com_Texture.cpp" />
     <ClCompile Include="Theatre.cpp" />
     <ClCompile Include="Transform.cpp" />
   </ItemGroup>
